Added va_copy to standard.h.

+++ src/Headers/cstringTable.h
@@ -54,7 +54,7 @@ extern /*@falsewhennull@*/ bool cstringTable_isDefined(cstringTable) /*@*/ ;
 extern /*@nullwhentrue@*/ /*@unused@*/ bool cstringTable_isUndefined(cstringTable) /*@*/ ;
 # define cstringTable_isUndefined(p_h) ((p_h) == cstringTable_undefined)
 
-extern /*@only@*/ cstringTable cstringTable_create(int p_size) /*@*/ ;
+extern /*@only@*/ cstringTable cstringTable_create(unsigned long p_size) /*@*/ ;
 extern void cstringTable_insert (cstringTable p_h, 
 				 /*@only@*/ cstring p_key, 
 				 int p_value) /*@modifies p_h@*/ ;
